# openapi.yaml
openapi: 3.0.3
info:
  title: API GreenFit - Sistema de Catraca de Academia
  description: API para gerenciamento de usuários da academia com controle de acesso por catraca
  version: 1.0.0
  contact:
    name: Suporte GreenFit
    email: suporte@greenfit.com

servers:
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento
  - url: https://seu-dominio.vercel.app
    description: Servidor de Produção

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do login

  schemas:
    Usuario:
      type: object
      required:
        - nome
        - cpf
      properties:
        nome:
          type: string
          description: Nome completo do usuário
          example: "João Silva"
          minLength: 3
          maxLength: 100
        cpf:
          type: string
          description: CPF do usuário (11 dígitos, apenas números)
          pattern: '^[0-9]{11}$'
          example: "12345678901"
          minLength: 11
          maxLength: 11
    
    UsuarioResponse:
      type: object
      properties:
        nome:
          type: string
          example: "João Silva"
        cpf:
          type: string
          example: "12345678901"
    
    ErrorResponse:
      type: object
      properties:
        erro:
          type: string
          description: Mensagem de erro
          example: "CPF inválido"
        error:
          type: string
          description: Mensagem de erro (para alguns casos)
          example: "URL não encontrada"
    
    SuccessResponse:
      type: object
      properties:
        mensagem:
          type: string
          description: Mensagem de sucesso
          example: "Usuário cadastrado com sucesso!"
    
    LoginRequest:
      type: object
      required:
        - usuario
        - senha
      properties:
        usuario:
          type: string
          description: Nome de usuário administrador
          example: "admin"
        senha:
          type: string
          description: Senha do administrador
          example: "admin123"
    
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Token JWT para autenticação
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        mensagem:
          type: string
          example: "Login realizado com sucesso!"

paths:
  /:
    get:
      summary: Rota principal da API
      description: Retorna uma mensagem de boas-vindas ao sistema GreenFit
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: "Bem-vindo ao GreenFit!"
        '401':
          description: Não autorizado - Token inválido ou ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /login:
    post:
      summary: Login do administrador
      description: Autentica o administrador e retorna um token JWT para usar nas demais rotas
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Dados não enviados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cadastro:
    post:
      summary: Cadastrar novo usuário
      description: Cadastra um novo usuário no sistema. Não requer autenticação.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
      responses:
        '201':
          description: Usuário cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Dados inválidos ou faltantes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: CPF já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /consulta:
    get:
      summary: Listar todos os usuários
      description: Retorna uma lista com todos os usuários cadastrados no sistema
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: object
                properties:
                  usuarios:
                    type: array
                    items:
                      $ref: '#/components/schemas/UsuarioResponse'
        '401':
          description: Não autorizado - Token inválido ou ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /consulta/{cpf}:
    get:
      summary: Consultar usuário por CPF
      description: Retorna os dados de um usuário específico baseado no CPF
      security:
        - BearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]{11}$'
          description: CPF do usuário (11 dígitos)
          example: "12345678901"
      responses:
        '200':
          description: Dados do usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /editar/{cpf}:
    patch:
      summary: Editar usuário parcialmente
      description: Atualiza campos específicos de um usuário (apenas o nome pode ser alterado)
      security:
        - BearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]{11}$'
          description: CPF do usuário a ser editado
          example: "12345678901"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  description: Novo nome do usuário
                  example: "João Silva Santos"
      responses:
        '200':
          description: Usuário editado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Dados inválidos ou tentativa de alterar CPF
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /substituir/{cpf}:
    put:
      summary: Substituir usuário completamente
      description: Substitui todos os dados de um usuário (substituição total)
      security:
        - BearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]{11}$'
          description: CPF do usuário a ser substituído
          example: "12345678901"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
              properties:
                nome:
                  type: string
                  description: Novo nome do usuário
                  example: "João Silva Santos"
      responses:
        '200':
          description: Usuário substituído com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Dados inválidos ou campo nome ausente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /excluir/{cpf}:
    delete:
      summary: Excluir usuário
      description: Remove um usuário do sistema permanentemente
      security:
        - BearerAuth: []
      parameters:
        - name: cpf
          in: path
          required: true
          schema:
            type: string
            pattern: '^[0-9]{11}$'
          description: CPF do usuário a ser excluído
          example: "12345678901"
      responses:
        '200':
          description: Usuário excluído com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: CPF inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

tags:
  - name: Público
    description: Rotas que não requerem autenticação
  - name: Protegidas
    description: Rotas que requerem token de autenticação